backend/requirements.txt
Flask
flask-cors
flask-limiter
spacy
emoji
fuzzywuzzy
python-Levenshtein
rapidfuzz
nltk
sentence-transformers
numpy
pytz
textblob